change the structure of project by using blueprints
