+++ flaskblog/templates/register.html
@@ -65,7 +65,7 @@
 </div>
 <div class="border-top pt-4">
     <small class="text-muted">
-        Already Have An Account? <a class="ml-2" href="{{ url_for('login') }}">Sign In</a>
+        Already Have An Account? <a class="ml-2" href="{{ url_for('users.login') }}">Sign In</a>
     </small>
 </div>
 {% endblock content %}
